feat: darkmode for activity diagram and nodetoolbar reusable
