+++ library/lib/components/svgs/index.ts
@@ -1,2 +1,3 @@
 export * from "./nodes"
 export * from "./edges"
+export * from "./StyledElements"
