+++ library/lib/nodes/activityDiagram/ActivityInitialNode.tsx
@@ -1,18 +1,11 @@
-import { NodeProps, NodeToolbar, Position, type Node } from "@xyflow/react"
+import { NodeProps, type Node } from "@xyflow/react"
 import { DefaultNodeWrapper, HandleId } from "../wrappers"
-
 import { DefaultNodeProps } from "@/types"
-import Box from "@mui/material/Box"
-import DeleteOutlineOutlinedIcon from "@mui/icons-material/DeleteOutlineOutlined"
-import EditIcon from "@mui/icons-material/Edit"
 import { useRef } from "react"
-import { usePopoverStore } from "@/store/context"
-import { useShallow } from "zustand/shallow"
-import { useHandleDelete } from "@/hooks/useHandleDelete"
 import { PopoverManager } from "@/components/popovers/PopoverManager"
 import { useDiagramModifiable } from "@/hooks/useDiagramModifiable"
-import { useIsOnlyThisElementSelected } from "@/hooks/useIsOnlyThisElementSelected"
 import { ActivityInitialNodeSVG } from "@/components/svgs/nodes"
+import { NodeToolbar } from "@/components/toolbars/NodeToolbar"
 
 export function ActivityInitialNode({
   id,
@@ -22,11 +15,6 @@ export function ActivityInitialNode({
   const svgWrapperRef = useRef<HTMLDivElement | null>(null)
 
   const isDiagramModifiable = useDiagramModifiable()
-  const selected = useIsOnlyThisElementSelected(id)
-  const setPopOverElementId = usePopoverStore(
-    useShallow((state) => state.setPopOverElementId)
-  )
-  const handleDelete = useHandleDelete(id)
 
   if (!width || !height) {
     return null
@@ -47,27 +35,7 @@ export function ActivityInitialNode({
         HandleId.LeftTop,
       ]}
     >
-      <NodeToolbar
-        isVisible={isDiagramModifiable && !!selected}
-        position={Position.Top}
-        align="end"
-        offset={10}
-      >
-        <Box sx={{ display: "flex", gap: 1, flexDirection: "column" }}>
-          <DeleteOutlineOutlinedIcon
-            onClick={handleDelete}
-            style={{ cursor: "pointer", width: 16, height: 16 }}
-          />
-
-          <EditIcon
-            onClick={() => {
-              setPopOverElementId(id)
-            }}
-            style={{ cursor: "pointer", width: 16, height: 16 }}
-          />
-        </Box>
-      </NodeToolbar>
-
+      <NodeToolbar elementId={id} />
       <div ref={svgWrapperRef}>
         <ActivityInitialNodeSVG
           width={width}
