+++ library/lib/nodes/classDiagram/Class.tsx
@@ -1,16 +1,9 @@
-import {
-  NodeProps,
-  NodeResizer,
-  NodeToolbar,
-  Position,
-  type Node,
-} from "@xyflow/react"
+import { NodeProps, NodeResizer, type Node } from "@xyflow/react"
 import { DefaultNodeWrapper } from "@/nodes/wrappers"
 import { ClassSVG } from "@/components"
 import { useEffect, useMemo, useRef } from "react"
-import { Box } from "@mui/material"
 import { ClassNodeProps } from "@/types"
-import { useDiagramStore, usePopoverStore } from "@/store/context"
+import { useDiagramStore } from "@/store/context"
 import { useShallow } from "zustand/shallow"
 import {
   measureTextWidth,
@@ -25,11 +18,10 @@ import {
   DEFAULT_HEADER_HEIGHT,
   DEFAULT_HEADER_HEIGHT_WITH_STREOTYPE,
 } from "@/constants"
-import { useHandleDelete } from "@/hooks/useHandleDelete"
 import { PopoverManager } from "@/components/popovers/PopoverManager"
 import { useDiagramModifiable } from "@/hooks/useDiagramModifiable"
 import { useIsOnlyThisElementSelected } from "@/hooks/useIsOnlyThisElementSelected"
-import { DeleteIcon, EditIcon } from "@/components/Icon"
+import { NodeToolbar } from "@/components/toolbars/NodeToolbar"
 
 export function Class({
   id,
@@ -42,14 +34,11 @@ export function Class({
       setNodes: state.setNodes,
     }))
   )
-  const setPopOverElementId = usePopoverStore(
-    useShallow((state) => state.setPopOverElementId)
-  )
+
   const selected = useIsOnlyThisElementSelected(id)
   const isDiagramModifiable = useDiagramModifiable()
 
   const classSvgWrapperRef = useRef<HTMLDivElement | null>(null)
-  const handleDelete = useHandleDelete(id)
 
   const showStereotype = !!stereotype
   const headerHeight = showStereotype
@@ -163,26 +152,7 @@ export function Class({
         maxHeight={minHeight}
         handleStyle={{ width: 8, height: 8 }}
       />
-      <NodeToolbar
-        isVisible={isDiagramModifiable && !!selected}
-        position={Position.Top}
-        align="end"
-        offset={10}
-      >
-        <Box sx={{ display: "flex", gap: 1, flexDirection: "column" }}>
-          <DeleteIcon
-            onClick={handleDelete}
-            style={{ cursor: "pointer", width: 16, height: 16 }}
-          />
-          <EditIcon
-            onClick={(e) => {
-              e.stopPropagation()
-              setPopOverElementId(id)
-            }}
-            style={{ cursor: "pointer", width: 16, height: 16 }}
-          />
-        </Box>
-      </NodeToolbar>
+      <NodeToolbar elementId={id} />
 
       <div ref={classSvgWrapperRef}>
         <ClassSVG
