+++ library/lib/nodes/classDiagram/ColorDescription.tsx
@@ -1,6 +1,7 @@
 import { NodeProps, type Node } from "@xyflow/react"
 import { DefaultNodeWrapper } from "../wrappers"
 import { ColorDescriptionSVG } from "@/components"
+import { NodeToolbar } from "@/components/toolbars/NodeToolbar"
 
 type Props = Node<{
   description: string
@@ -18,6 +19,8 @@ export function ColorDescription({
 
   return (
     <DefaultNodeWrapper width={width} height={height} elementId={id}>
+      <NodeToolbar elementId={id} />
+
       <ColorDescriptionSVG
         description={description}
         width={width}
