+++ library/lib/nodes/componentDiagram/ComponentInterface.tsx
@@ -1,17 +1,11 @@
-import { NodeProps, NodeToolbar, Position, type Node } from "@xyflow/react"
+import { NodeProps, type Node } from "@xyflow/react"
 import { DefaultNodeWrapper, HandleId } from "../wrappers"
 import { DefaultNodeProps } from "@/types"
-import Box from "@mui/material/Box"
-import DeleteOutlineOutlinedIcon from "@mui/icons-material/DeleteOutlineOutlined"
-import EditIcon from "@mui/icons-material/Edit"
 import { useRef } from "react"
-import { usePopoverStore } from "@/store/context"
-import { useShallow } from "zustand/shallow"
-import { useHandleDelete } from "@/hooks/useHandleDelete"
 import { PopoverManager } from "@/components/popovers/PopoverManager"
 import { useDiagramModifiable } from "@/hooks/useDiagramModifiable"
-import { useIsOnlyThisElementSelected } from "@/hooks/useIsOnlyThisElementSelected"
 import { ComponentInterfaceNodeSVG } from "@/components"
+import { NodeToolbar } from "@/components/toolbars/NodeToolbar"
 
 export function ComponentInterface({
   id,
@@ -20,13 +14,7 @@ export function ComponentInterface({
   data: { name },
 }: NodeProps<Node<DefaultNodeProps>>) {
   const svgWrapperRef = useRef<HTMLDivElement | null>(null)
-
   const isDiagramModifiable = useDiagramModifiable()
-  const selected = useIsOnlyThisElementSelected(id)
-  const setPopOverElementId = usePopoverStore(
-    useShallow((state) => state.setPopOverElementId)
-  )
-  const handleDelete = useHandleDelete(id)
 
   if (!width || !height) {
     return null
@@ -48,26 +36,8 @@ export function ComponentInterface({
         HandleId.LeftTop,
       ]}
     >
-      <NodeToolbar
-        isVisible={isDiagramModifiable && !!selected}
-        position={Position.Top}
-        align="end"
-        offset={10}
-      >
-        <Box sx={{ display: "flex", gap: 1, flexDirection: "column" }}>
-          <DeleteOutlineOutlinedIcon
-            onClick={handleDelete}
-            style={{ cursor: "pointer", width: 16, height: 16 }}
-          />
+      <NodeToolbar elementId={id} />
 
-          <EditIcon
-            onClick={() => {
-              setPopOverElementId(id)
-            }}
-            style={{ cursor: "pointer", width: 16, height: 16 }}
-          />
-        </Box>
-      </NodeToolbar>
       <div ref={svgWrapperRef}>
         <ComponentInterfaceNodeSVG
           width={width}
