+++ library/lib/nodes/flowchart/FlowchartFunctionCall.tsx
@@ -1,24 +1,13 @@
-import {
-  NodeProps,
-  NodeResizer,
-  NodeToolbar,
-  Position,
-  type Node,
-} from "@xyflow/react"
+import { NodeProps, NodeResizer, type Node } from "@xyflow/react"
 import { DefaultNodeWrapper } from "../wrappers"
 import { useHandleOnResize } from "@/hooks"
 import { DefaultNodeProps } from "@/types"
-import Box from "@mui/material/Box"
-import DeleteOutlineOutlinedIcon from "@mui/icons-material/DeleteOutlineOutlined"
-import EditIcon from "@mui/icons-material/Edit"
 import { useRef } from "react"
-import { usePopoverStore } from "@/store/context"
-import { useShallow } from "zustand/shallow"
-import { useHandleDelete } from "@/hooks/useHandleDelete"
 import { PopoverManager } from "@/components/popovers/PopoverManager"
 import { useDiagramModifiable } from "@/hooks/useDiagramModifiable"
 import { useIsOnlyThisElementSelected } from "@/hooks/useIsOnlyThisElementSelected"
 import { FlowchartFunctionCallNodeSVG } from "@/components"
+import { NodeToolbar } from "@/components/toolbars/NodeToolbar"
 
 export function FlowchartFunctionCall({
   id,
@@ -31,10 +20,6 @@ export function FlowchartFunctionCall({
   const { onResize } = useHandleOnResize(parentId)
   const isDiagramModifiable = useDiagramModifiable()
   const selected = useIsOnlyThisElementSelected(id)
-  const setPopOverElementId = usePopoverStore(
-    useShallow((state) => state.setPopOverElementId)
-  )
-  const handleDelete = useHandleDelete(id)
 
   if (!width || !height) {
     return null
@@ -42,26 +27,7 @@ export function FlowchartFunctionCall({
 
   return (
     <DefaultNodeWrapper width={width} height={height} elementId={id}>
-      <NodeToolbar
-        isVisible={isDiagramModifiable && !!selected}
-        position={Position.Top}
-        align="end"
-        offset={10}
-      >
-        <Box sx={{ display: "flex", gap: 1, flexDirection: "column" }}>
-          <DeleteOutlineOutlinedIcon
-            onClick={handleDelete}
-            style={{ cursor: "pointer", width: 16, height: 16 }}
-          />
-
-          <EditIcon
-            onClick={() => {
-              setPopOverElementId(id)
-            }}
-            style={{ cursor: "pointer", width: 16, height: 16 }}
-          />
-        </Box>
-      </NodeToolbar>
+      <NodeToolbar elementId={id} />
       <NodeResizer
         isVisible={isDiagramModifiable && !!selected}
         onResize={onResize}
