+++ library/lib/nodes/objectDiagram/ObjectName.tsx
@@ -1,18 +1,9 @@
-import {
-  NodeProps,
-  NodeResizer,
-  NodeToolbar,
-  Position,
-  type Node,
-} from "@xyflow/react"
+import { NodeProps, NodeResizer, type Node } from "@xyflow/react"
 import { DefaultNodeWrapper } from "@/nodes/wrappers"
 import { ObjectNameSVG } from "@/components"
-import EditIcon from "@mui/icons-material/Edit"
-import DeleteOutlineOutlinedIcon from "@mui/icons-material/DeleteOutlineOutlined"
 import { useEffect, useMemo, useRef } from "react"
-import { Box } from "@mui/material"
 import { ObjectNodeProps } from "@/types"
-import { useDiagramStore, usePopoverStore } from "@/store/context"
+import { useDiagramStore } from "@/store/context"
 import { useShallow } from "zustand/shallow"
 import {
   measureTextWidth,
@@ -26,10 +17,10 @@ import {
   DEFAULT_FONT,
   DEFAULT_HEADER_HEIGHT,
 } from "@/constants"
-import { useHandleDelete } from "@/hooks/useHandleDelete"
 import { PopoverManager } from "@/components/popovers/PopoverManager"
 import { useDiagramModifiable } from "@/hooks/useDiagramModifiable"
 import { useIsOnlyThisElementSelected } from "@/hooks/useIsOnlyThisElementSelected"
+import { NodeToolbar } from "@/components/toolbars/NodeToolbar"
 
 export function ObjectName({
   id,
@@ -42,14 +33,11 @@ export function ObjectName({
       setNodes: state.setNodes,
     }))
   )
-  const setPopOverElementId = usePopoverStore(
-    useShallow((state) => state.setPopOverElementId)
-  )
+
   const selected = useIsOnlyThisElementSelected(id)
   const isDiagramModifiable = useDiagramModifiable()
 
   const objectSvgWrapperRef = useRef<HTMLDivElement | null>(null)
-  const handleDelete = useHandleDelete(id)
 
   // Object diagrams don't have stereotypes, so header height is consistent
   const headerHeight = DEFAULT_HEADER_HEIGHT
@@ -149,6 +137,7 @@ export function ObjectName({
       elementId={id}
       className="horizontally-not-resizable"
     >
+      <NodeToolbar elementId={id} />
       <NodeResizer
         nodeId={id}
         isVisible={isDiagramModifiable && !!selected}
@@ -157,26 +146,6 @@ export function ObjectName({
         maxHeight={minHeight}
         handleStyle={{ width: 8, height: 8 }}
       />
-      <NodeToolbar
-        isVisible={isDiagramModifiable && !!selected}
-        position={Position.Top}
-        align="end"
-        offset={10}
-      >
-        <Box sx={{ display: "flex", gap: 1, flexDirection: "column" }}>
-          <DeleteOutlineOutlinedIcon
-            onClick={handleDelete}
-            style={{ cursor: "pointer", width: 16, height: 16 }}
-          />
-          <EditIcon
-            onClick={(e) => {
-              e.stopPropagation()
-              setPopOverElementId(id)
-            }}
-            style={{ cursor: "pointer", width: 16, height: 16 }}
-          />
-        </Box>
-      </NodeToolbar>
 
       <div ref={objectSvgWrapperRef}>
         <ObjectNameSVG
