+++ library/lib/nodes/petriNetDiagram/PetriNetPlace.tsx
@@ -1,17 +1,11 @@
-import { NodeProps, NodeToolbar, Position, type Node } from "@xyflow/react"
+import { NodeProps, type Node } from "@xyflow/react"
 import { PetriNetPlaceSVG } from "@/components"
-import EditIcon from "@mui/icons-material/Edit"
-import DeleteOutlineOutlinedIcon from "@mui/icons-material/DeleteOutlineOutlined"
 import { useRef } from "react"
-import { Box } from "@mui/material"
 import { PetriNetPlaceProps } from "@/types"
-import { usePopoverStore } from "@/store/context"
-import { useShallow } from "zustand/shallow"
-import { useHandleDelete } from "@/hooks/useHandleDelete"
 import { PopoverManager } from "@/components/popovers/PopoverManager"
 import { useDiagramModifiable } from "@/hooks/useDiagramModifiable"
-import { useIsOnlyThisElementSelected } from "@/hooks/useIsOnlyThisElementSelected"
 import { DefaultNodeWrapper, HandleId } from "../wrappers"
+import { NodeToolbar } from "@/components/toolbars/NodeToolbar"
 
 export function PetriNetPlace({
   id,
@@ -21,11 +15,6 @@ export function PetriNetPlace({
 }: NodeProps<Node<PetriNetPlaceProps>>) {
   const svgWrapperRef = useRef<HTMLDivElement | null>(null)
   const isDiagramModifiable = useDiagramModifiable()
-  const selected = useIsOnlyThisElementSelected(id)
-  const setPopOverElementId = usePopoverStore(
-    useShallow((state) => state.setPopOverElementId)
-  )
-  const handleDelete = useHandleDelete(id)
 
   if (!width || !height) {
     return null
@@ -47,27 +36,7 @@ export function PetriNetPlace({
         HandleId.LeftTop,
       ]}
     >
-      <NodeToolbar
-        isVisible={isDiagramModifiable && !!selected}
-        position={Position.Top}
-        align="end"
-        offset={10}
-      >
-        <Box sx={{ display: "flex", gap: 1, flexDirection: "column" }}>
-          <DeleteOutlineOutlinedIcon
-            onClick={handleDelete}
-            style={{ cursor: "pointer", width: 16, height: 16 }}
-          />
-
-          <EditIcon
-            onClick={() => {
-              setPopOverElementId(id)
-            }}
-            style={{ cursor: "pointer", width: 16, height: 16 }}
-          />
-        </Box>
-      </NodeToolbar>
-
+      <NodeToolbar elementId={id} />
       <div ref={svgWrapperRef}>
         <PetriNetPlaceSVG
           width={width}
