+++ library/lib/nodes/sfcDiagram/SfcTransitionBranch.tsx
@@ -1,13 +1,9 @@
-import { NodeProps, NodeToolbar, Position, type Node } from "@xyflow/react"
+import { NodeProps, type Node } from "@xyflow/react"
 import { DefaultNodeWrapper } from "../wrappers"
-import Box from "@mui/material/Box"
-import DeleteOutlineOutlinedIcon from "@mui/icons-material/DeleteOutlineOutlined"
 import { useRef } from "react"
-import { useHandleDelete } from "@/hooks/useHandleDelete"
-import { useDiagramModifiable } from "@/hooks/useDiagramModifiable"
-import { useIsOnlyThisElementSelected } from "@/hooks/useIsOnlyThisElementSelected"
 import { DefaultNodeProps } from "@/types"
 import { SfcTransitionBranchNodeSVG } from "@/components"
+import { NodeToolbar } from "@/components/toolbars/NodeToolbar"
 
 export function SfcTransitionBranch({
   id,
@@ -16,9 +12,6 @@ export function SfcTransitionBranch({
   data: { name },
 }: NodeProps<Node<DefaultNodeProps>>) {
   const svgWrapperRef = useRef<HTMLDivElement | null>(null)
-  const isDiagramModifiable = useDiagramModifiable()
-  const selected = useIsOnlyThisElementSelected(id)
-  const handleDelete = useHandleDelete(id)
 
   if (!width || !height) {
     return null
@@ -26,19 +19,7 @@ export function SfcTransitionBranch({
 
   return (
     <DefaultNodeWrapper width={width} height={height} elementId={id}>
-      <NodeToolbar
-        isVisible={isDiagramModifiable && !!selected}
-        position={Position.Top}
-        align="end"
-        offset={10}
-      >
-        <Box sx={{ display: "flex", gap: 1, flexDirection: "column" }}>
-          <DeleteOutlineOutlinedIcon
-            onClick={handleDelete}
-            style={{ cursor: "pointer", width: 16, height: 16 }}
-          />
-        </Box>
-      </NodeToolbar>
+      <NodeToolbar elementId={id} />
       <div ref={svgWrapperRef}>
         <SfcTransitionBranchNodeSVG
           width={width}
